@@ -1,19 +1,21 @@
 module github.com/conduitio-labs/conduit-connector-cosmos-nosql
 
-go 1.22.2
+go 1.23.2
 
 require (
 	github.com/Azure/azure-sdk-for-go/sdk/data/azcosmos v1.1.0
-	github.com/conduitio/conduit-connector-sdk v0.9.1
-	github.com/golang/mock v1.6.0
+	github.com/conduitio/conduit-commons v0.4.0
+	github.com/conduitio/conduit-connector-sdk v0.11.0
 	github.com/golangci/golangci-lint v1.61.0
 	github.com/huandu/go-sqlbuilder v1.30.1
 	github.com/matryer/is v1.4.1
+	go.uber.org/mock v0.4.0
 )
 
 require (
 	4d63.com/gocheckcompilerdirectives v1.2.1 // indirect
 	4d63.com/gochecknoglobals v0.2.1 // indirect
+	dario.cat/mergo v1.0.1 // indirect
 	github.com/4meepo/tagalign v1.3.4 // indirect
 	github.com/Abirdcfly/dupword v0.1.1 // indirect
 	github.com/Antonboom/errname v0.1.13 // indirect
