@@ -1,5 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
-// Source: source/source.go
+// Source: github.com/conduitio-labs/conduit-connector-cosmos-nosql/source (interfaces: Iterator)
+//
+// Generated by this command:
+//
+//	mockgen -typed -destination mock/source.go -package mock . Iterator
+//
 
 // Package mock is a generated GoMock package.
 package mock
@@ -8,8 +13,8 @@ import (
 	context "context"
 	reflect "reflect"
 
-	sdk "github.com/conduitio/conduit-connector-sdk"
-	gomock "github.com/golang/mock/gomock"
+	opencdc "github.com/conduitio/conduit-commons/opencdc"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockIterator is a mock of Iterator interface.
@@ -45,22 +50,70 @@ func (m *MockIterator) HasNext(arg0 context.Context) (bool, error) {
 }
 
 // HasNext indicates an expected call of HasNext.
-func (mr *MockIteratorMockRecorder) HasNext(arg0 interface{}) *gomock.Call {
+func (mr *MockIteratorMockRecorder) HasNext(arg0 any) *MockIteratorHasNextCall {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "HasNext", reflect.TypeOf((*MockIterator)(nil).HasNext), arg0)
+	call := mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "HasNext", reflect.TypeOf((*MockIterator)(nil).HasNext), arg0)
+	return &MockIteratorHasNextCall{Call: call}
+}
+
+// MockIteratorHasNextCall wrap *gomock.Call
+type MockIteratorHasNextCall struct {
+	*gomock.Call
+}
+
+// Return rewrite *gomock.Call.Return
+func (c *MockIteratorHasNextCall) Return(arg0 bool, arg1 error) *MockIteratorHasNextCall {
+	c.Call = c.Call.Return(arg0, arg1)
+	return c
+}
+
+// Do rewrite *gomock.Call.Do
+func (c *MockIteratorHasNextCall) Do(f func(context.Context) (bool, error)) *MockIteratorHasNextCall {
+	c.Call = c.Call.Do(f)
+	return c
+}
+
+// DoAndReturn rewrite *gomock.Call.DoAndReturn
+func (c *MockIteratorHasNextCall) DoAndReturn(f func(context.Context) (bool, error)) *MockIteratorHasNextCall {
+	c.Call = c.Call.DoAndReturn(f)
+	return c
 }
 
 // Next mocks base method.
-func (m *MockIterator) Next() (sdk.Record, error) {
+func (m *MockIterator) Next() (opencdc.Record, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "Next")
-	ret0, _ := ret[0].(sdk.Record)
+	ret0, _ := ret[0].(opencdc.Record)
 	ret1, _ := ret[1].(error)
 	return ret0, ret1
 }
 
 // Next indicates an expected call of Next.
-func (mr *MockIteratorMockRecorder) Next() *gomock.Call {
+func (mr *MockIteratorMockRecorder) Next() *MockIteratorNextCall {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Next", reflect.TypeOf((*MockIterator)(nil).Next))
+	call := mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Next", reflect.TypeOf((*MockIterator)(nil).Next))
+	return &MockIteratorNextCall{Call: call}
+}
+
+// MockIteratorNextCall wrap *gomock.Call
+type MockIteratorNextCall struct {
+	*gomock.Call
+}
+
+// Return rewrite *gomock.Call.Return
+func (c *MockIteratorNextCall) Return(arg0 opencdc.Record, arg1 error) *MockIteratorNextCall {
+	c.Call = c.Call.Return(arg0, arg1)
+	return c
+}
+
+// Do rewrite *gomock.Call.Do
+func (c *MockIteratorNextCall) Do(f func() (opencdc.Record, error)) *MockIteratorNextCall {
+	c.Call = c.Call.Do(f)
+	return c
+}
+
+// DoAndReturn rewrite *gomock.Call.DoAndReturn
+func (c *MockIteratorNextCall) DoAndReturn(f func() (opencdc.Record, error)) *MockIteratorNextCall {
+	c.Call = c.Call.DoAndReturn(f)
+	return c
 }
