@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+//go:generate mockgen -typed -destination mock/source.go -package mock . Iterator
+
 package source
 
 import (
@@ -19,13 +21,15 @@ import (
 	"fmt"
 
 	"github.com/conduitio-labs/conduit-connector-cosmos-nosql/common"
+	"github.com/conduitio/conduit-commons/config"
+	"github.com/conduitio/conduit-commons/opencdc"
 	sdk "github.com/conduitio/conduit-connector-sdk"
 )
 
 // Iterator is an interface needed for the [Source].
 type Iterator interface {
 	HasNext(context.Context) (bool, error)
-	Next() (sdk.Record, error)
+	Next() (opencdc.Record, error)
 }
 
 // Source is an Azure Cosmos DB for NoSQL source connector.
@@ -41,16 +45,16 @@ func New() sdk.Source {
 	return sdk.SourceWithMiddleware(&Source{}, sdk.DefaultSourceMiddleware()...)
 }
 
-// Parameters is a map of named [sdk.Parameter] that describe how to configure the [Source].
-func (s *Source) Parameters() map[string]sdk.Parameter {
+// Parameters is a map of named [config.Parameter] that describe how to configure the [Source].
+func (s *Source) Parameters() config.Parameters {
 	return s.config.Parameters()
 }
 
 // Configure parses and initializes the [Source] config.
-func (s *Source) Configure(ctx context.Context, raw map[string]string) error {
+func (s *Source) Configure(ctx context.Context, raw config.Config) error {
 	sdk.Logger(ctx).Info().Msg("Configuring an Azure Cosmos DB for NoSQL Source...")
 
-	if err := sdk.Util.ParseConfig(raw, &s.config); err != nil {
+	if err := sdk.Util.ParseConfig(ctx, raw, &s.config, New().Parameters()); err != nil {
 		return fmt.Errorf("parse source config: %w", err)
 	}
 
@@ -63,7 +67,7 @@ func (s *Source) Configure(ctx context.Context, raw map[string]string) error {
 }
 
 // Open parses the position and initializes the iterator.
-func (s *Source) Open(ctx context.Context, sdkPosition sdk.Position) error {
+func (s *Source) Open(ctx context.Context, sdkPosition opencdc.Position) error {
 	sdk.Logger(ctx).Info().Msg("Opening an Azure Cosmos DB for NoSQL Source...")
 
 	var err error
@@ -76,29 +80,29 @@ func (s *Source) Open(ctx context.Context, sdkPosition sdk.Position) error {
 	return nil
 }
 
-// Read returns the next [sdk.Record].
-func (s *Source) Read(ctx context.Context) (sdk.Record, error) {
+// Read returns the next [opencdc.Record].
+func (s *Source) Read(ctx context.Context) (opencdc.Record, error) {
 	sdk.Logger(ctx).Debug().Msg("Reading a record from an Azure Cosmos DB for NoSQL Source...")
 
 	hasNext, err := s.iterator.HasNext(ctx)
 	if err != nil {
-		return sdk.Record{}, fmt.Errorf("has next: %w", err)
+		return opencdc.Record{}, fmt.Errorf("has next: %w", err)
 	}
 
 	if !hasNext {
-		return sdk.Record{}, sdk.ErrBackoffRetry
+		return opencdc.Record{}, sdk.ErrBackoffRetry
 	}
 
 	record, err := s.iterator.Next()
 	if err != nil {
-		return sdk.Record{}, fmt.Errorf("next: %w", err)
+		return opencdc.Record{}, fmt.Errorf("next: %w", err)
 	}
 
 	return record, nil
 }
 
 // Ack just logs the debug event with the position.
-func (s *Source) Ack(ctx context.Context, sdkPosition sdk.Position) error {
+func (s *Source) Ack(ctx context.Context, sdkPosition opencdc.Position) error {
 	sdk.Logger(ctx).Debug().Str("position", string(sdkPosition)).Msg("got ack")
 
 	return nil
