@@ -1,87 +1,100 @@
 // Code generated by paramgen. DO NOT EDIT.
-// Source: github.com/conduitio/conduit-connector-sdk/cmd/paramgen
+// Source: github.com/ConduitIO/conduit-commons/tree/main/paramgen
 
 package source
 
 import (
-	sdk "github.com/conduitio/conduit-connector-sdk"
+	"github.com/conduitio/conduit-commons/config"
 )
 
-func (Config) Parameters() map[string]sdk.Parameter {
-	return map[string]sdk.Parameter{
-		"batchSize": {
+const (
+	ConfigBatchSize      = "batchSize"
+	ConfigContainer      = "container"
+	ConfigDatabase       = "database"
+	ConfigKeys           = "keys"
+	ConfigMetaProperties = "metaProperties"
+	ConfigOrderingKey    = "orderingKey"
+	ConfigPartitionValue = "partitionValue"
+	ConfigPrimaryKey     = "primaryKey"
+	ConfigSnapshot       = "snapshot"
+	ConfigUri            = "uri"
+)
+
+func (Config) Parameters() map[string]config.Parameter {
+	return map[string]config.Parameter{
+		ConfigBatchSize: {
 			Default:     "1000",
 			Description: "The size of an element batch.",
-			Type:        sdk.ParameterTypeInt,
-			Validations: []sdk.Validation{
-				sdk.ValidationGreaterThan{Value: 0},
-				sdk.ValidationLessThan{Value: 100001},
+			Type:        config.ParameterTypeInt,
+			Validations: []config.Validation{
+				config.ValidationGreaterThan{V: 0},
+				config.ValidationLessThan{V: 100001},
 			},
 		},
-		"container": {
+		ConfigContainer: {
 			Default:     "",
 			Description: "The name of a container the connector should work with.",
-			Type:        sdk.ParameterTypeString,
-			Validations: []sdk.Validation{
-				sdk.ValidationRequired{},
+			Type:        config.ParameterTypeString,
+			Validations: []config.Validation{
+				config.ValidationRequired{},
 			},
 		},
-		"database": {
+		ConfigDatabase: {
 			Default:     "",
 			Description: "The name of a database the connector should work with.",
-			Type:        sdk.ParameterTypeString,
-			Validations: []sdk.Validation{
-				sdk.ValidationRequired{},
+			Type:        config.ParameterTypeString,
+			Validations: []config.Validation{
+				config.ValidationRequired{},
 			},
 		},
-		"keys": {
+		ConfigKeys: {
 			Default:     "",
-			Description: "Comma-separated list of key names to build the sdk.Record.Key.",
-			Type:        sdk.ParameterTypeString,
-			Validations: []sdk.Validation{},
+			Description: "Comma-separated list of key names to build the opencdc.Record.Key.",
+			Type:        config.ParameterTypeString,
+			Validations: []config.Validation{},
 		},
-		"metaProperties": {
+		ConfigMetaProperties: {
 			Default:     "false",
-			Description: "metaProperties whether the connector takes the next automatically generated meta-properties: \"_rid\", \"_ts\", \"_self\", \"_etag\", \"_attachments\".",
-			Type:        sdk.ParameterTypeBool,
-			Validations: []sdk.Validation{},
+			Description: "MetaProperties whether the connector takes\nthe next automatically generated meta-properties:\n\"_rid\", \"_ts\", \"_self\", \"_etag\", \"_attachments\".",
+			Type:        config.ParameterTypeBool,
+			Validations: []config.Validation{},
 		},
-		"orderingKey": {
+		ConfigOrderingKey: {
 			Default:     "",
 			Description: "The name of a key that is used for ordering items.",
-			Type:        sdk.ParameterTypeString,
-			Validations: []sdk.Validation{
-				sdk.ValidationRequired{},
+			Type:        config.ParameterTypeString,
+			Validations: []config.Validation{
+				config.ValidationRequired{},
 			},
 		},
-		"partitionValue": {
+		ConfigPartitionValue: {
 			Default:     "",
 			Description: "The logical partition key value.",
-			Type:        sdk.ParameterTypeString,
-			Validations: []sdk.Validation{
-				sdk.ValidationRequired{},
+			Type:        config.ParameterTypeString,
+			Validations: []config.Validation{
+				config.ValidationRequired{},
 			},
 		},
-		"primaryKey": {
+		ConfigPrimaryKey: {
 			Default:     "",
 			Description: "The key for authentication with Azure Cosmos DB.",
-			Type:        sdk.ParameterTypeString,
-			Validations: []sdk.Validation{
-				sdk.ValidationRequired{},
+			Type:        config.ParameterTypeString,
+			Validations: []config.Validation{
+				config.ValidationRequired{},
 			},
 		},
-		"snapshot": {
+		ConfigSnapshot: {
 			Default:     "true",
-			Description: "Determines whether the connector takes a snapshot of all items before starting CDC mode.",
-			Type:        sdk.ParameterTypeBool,
-			Validations: []sdk.Validation{},
+			Description: "Determines whether the connector takes a snapshot\nof all items before starting CDC mode.",
+			Type:        config.ParameterTypeBool,
+			Validations: []config.Validation{},
 		},
-		"uri": {
+		ConfigUri: {
 			Default:     "",
 			Description: "The connection uri pointed to an Azure Cosmos DB for NoSQL instance.",
-			Type:        sdk.ParameterTypeString,
-			Validations: []sdk.Validation{
-				sdk.ValidationRequired{},
+			Type:        config.ParameterTypeString,
+			Validations: []config.Validation{
+				config.ValidationRequired{},
 			},
 		},
 	}
