@@ -22,9 +22,9 @@ import (
 
 	"github.com/conduitio-labs/conduit-connector-cosmos-nosql/config"
 	"github.com/conduitio-labs/conduit-connector-cosmos-nosql/source/mock"
-	sdk "github.com/conduitio/conduit-connector-sdk"
-	"github.com/golang/mock/gomock"
+	"github.com/conduitio/conduit-commons/opencdc"
 	"github.com/matryer/is"
+	"go.uber.org/mock/gomock"
 )
 
 func TestSource_Configure_success(t *testing.T) {
@@ -77,17 +77,17 @@ func TestSource_Read_success(t *testing.T) {
 	ctrl := gomock.NewController(t)
 	ctx := context.Background()
 
-	key := make(sdk.StructuredData)
+	key := make(opencdc.StructuredData)
 	key["field1"] = 1
 
-	metadata := make(sdk.Metadata)
+	metadata := make(opencdc.Metadata)
 	metadata.SetCreatedAt(time.Time{})
 
-	record := sdk.Record{
-		Position: sdk.Position(`{"lastProcessedValue": 1}`),
+	record := opencdc.Record{
+		Position: opencdc.Position(`{"lastProcessedValue": 1}`),
 		Metadata: metadata,
 		Key:      key,
-		Payload: sdk.Change{
+		Payload: opencdc.Change{
 			After: key,
 		},
 	}
@@ -131,7 +131,7 @@ func TestSource_Read_failNext(t *testing.T) {
 
 	it := mock.NewMockIterator(ctrl)
 	it.EXPECT().HasNext(ctx).Return(true, nil)
-	it.EXPECT().Next().Return(sdk.Record{}, errors.New("marshal item: some error"))
+	it.EXPECT().Next().Return(opencdc.Record{}, errors.New("marshal item: some error"))
 
 	s := Source{iterator: it}
 
