@@ -17,6 +17,7 @@
 package main
 
 import (
-	_ "github.com/conduitio/conduit-connector-sdk/cmd/paramgen"
+	_ "github.com/conduitio/conduit-commons/paramgen"
 	_ "github.com/golangci/golangci-lint/cmd/golangci-lint"
+	_ "go.uber.org/mock/mockgen"
 )
